refactor request queue
In IE8, respond fails when requesting a file that we don't have a proxy
for which cause the remainder of the polyfill not to be parsed when the
first file requested is x-domain. If the first file is not x-domain,
this is not an issue because the files are processed in an async
waterfall after the first request. No other files will be processed if
this happens.

If the requests readyState is done immediately after invoking req.open
then the file will not be processed nor will the remainder of the files
in the queue (if any).

If the requests readyState is done and the response code is not 200 OK
or 304 Not Modified then the file will not be processed nor will the
remainder of the files in the queue (if any).

This fixes that and all tests pass.